refactor(scoreboard): merge visibility helpers and slim tests
