+++ fission/src/ui/components/topbar/GameplayControls.tsx
@@ -3,8 +3,7 @@ import type React from "react"
 import { useMemo } from "react"
 import { SCOREBOARD_MODES } from "@/systems/preferences/PreferenceTypes"
 import { useUIContext } from "@/ui/helpers/UIProviderHelpers"
-import { SCOREBOARD_GLYPH_SX, SCOREBOARD_MODE_LABELS } from "@/ui/helpers/ScoreboardVisibility"
-import { useScoreboard } from "@/ui/helpers/useScoreboard"
+import { SCOREBOARD_GLYPH_SX, SCOREBOARD_MODE_LABELS, useScoreboard } from "@/ui/helpers/ScoreboardVisibility"
 import MatchModeConfigPanel from "@/ui/panels/configuring/MatchModeConfigPanel"
 import MultiplayerStartModal from "@/ui/modals/MultiplayerStartModal"
 import { startMultiplayerWorld } from "@/ui/helpers/StartMultiplayerWorld"
